+++ test/core/utils/TransactionMapping.spec.ts
@@ -25,6 +25,7 @@ import { PropertyModificationType } from '../../../src/model/account/PropertyMod
 import { PropertyType } from '../../../src/model/account/PropertyType';
 import { PublicAccount } from '../../../src/model/account/PublicAccount';
 import { NetworkType } from '../../../src/model/blockchain/NetworkType';
+import { EncryptedMessage } from '../../../src/model/model';
 import { MosaicId } from '../../../src/model/mosaic/MosaicId';
 import { MosaicNonce } from '../../../src/model/mosaic/MosaicNonce';
 import { MosaicProperties } from '../../../src/model/mosaic/MosaicProperties';
@@ -34,6 +35,7 @@ import { AliasActionType } from '../../../src/model/namespace/AliasActionType';
 import { NamespaceId } from '../../../src/model/namespace/NamespaceId';
 import { NamespaceType } from '../../../src/model/namespace/NamespaceType';
 import { AccountLinkTransaction } from '../../../src/model/transaction/AccountLinkTransaction';
+import { AccountPropertyModification } from '../../../src/model/transaction/AccountPropertyModification';
 import { AccountPropertyTransaction } from '../../../src/model/transaction/AccountPropertyTransaction';
 import { AddressAliasTransaction } from '../../../src/model/transaction/AddressAliasTransaction';
 import { AggregateTransaction } from '../../../src/model/transaction/AggregateTransaction';
@@ -41,6 +43,7 @@ import { Deadline } from '../../../src/model/transaction/Deadline';
 import { HashType } from '../../../src/model/transaction/HashType';
 import { LinkAction } from '../../../src/model/transaction/LinkAction';
 import { LockFundsTransaction } from '../../../src/model/transaction/LockFundsTransaction';
+import { MessageType } from '../../../src/model/transaction/MessageType';
 import { ModifyAccountPropertyAddressTransaction } from '../../../src/model/transaction/ModifyAccountPropertyAddressTransaction';
 import { ModifyAccountPropertyMosaicTransaction } from '../../../src/model/transaction/ModifyAccountPropertyMosaicTransaction';
 import { ModifyMultisigAccountTransaction } from '../../../src/model/transaction/ModifyMultisigAccountTransaction';
@@ -67,7 +70,7 @@ describe('TransactionMapping - createFromPayload', () => {
 
     it('should create AccountPropertyAddressTransaction', () => {
         const address = Address.createFromRawAddress('SBILTA367K2LX2FEXG5TFWAS7GEFYAGY7QLFBYKC');
-        const addressPropertyFilter = AccountPropertyTransaction.createAddressFilter(
+        const addressPropertyFilter = AccountPropertyModification.createForAddress(
             PropertyModificationType.Add,
             address,
         );
@@ -89,7 +92,7 @@ describe('TransactionMapping - createFromPayload', () => {
 
     it('should create AccountPropertyMosaicTransaction', () => {
         const mosaicId = new MosaicId([2262289484, 3405110546]);
-        const mosaicPropertyFilter = AccountPropertyTransaction.createMosaicFilter(
+        const mosaicPropertyFilter = AccountPropertyModification.createForMosaic(
             PropertyModificationType.Add,
             mosaicId,
         );
@@ -111,7 +114,7 @@ describe('TransactionMapping - createFromPayload', () => {
 
     it('should create AccountPropertyMosaicTransaction', () => {
         const entityType = TransactionType.ADDRESS_ALIAS;
-        const entityTypePropertyFilter = AccountPropertyTransaction.createEntityTypeFilter(
+        const entityTypePropertyFilter = AccountPropertyModification.createForEntityType(
             PropertyModificationType.Add,
             entityType,
         );
@@ -572,6 +575,23 @@ describe('TransactionMapping - createFromDTO (Transaction.toJSON() feed)', () =>
         expect(transaction.message.payload).to.be.equal('test-message');
     });
 
+    it('should create TransferTransaction - Encrypted Message', () => {
+        const transferTransaction = TransferTransaction.create(
+            Deadline.create(),
+            Address.createFromRawAddress('SBILTA367K2LX2FEXG5TFWAS7GEFYAGY7QLFBYKC'),
+            [
+                NetworkCurrencyMosaic.createRelative(100),
+            ],
+            new EncryptedMessage('12324556'),
+            NetworkType.MIJIN_TEST,
+        );
+
+        const transaction = TransactionMapping.createFromDTO(transferTransaction.toJSON()) as TransferTransaction;
+
+        expect((transaction.recipient as Address).plain()).to.be.equal('SBILTA367K2LX2FEXG5TFWAS7GEFYAGY7QLFBYKC');
+        expect(transaction.message.type).to.be.equal(MessageType.EncryptedMessage);
+    });
+
     it('should create AccountLinkTransaction', () => {
         const accountLinkTransaction = AccountLinkTransaction.create(
             Deadline.create(),
@@ -588,7 +608,7 @@ describe('TransactionMapping - createFromDTO (Transaction.toJSON() feed)', () =>
 
     it('should create AccountPropertyAddressTransaction', () => {
         const address = Address.createFromRawAddress('SBILTA367K2LX2FEXG5TFWAS7GEFYAGY7QLFBYKC');
-        const addressPropertyFilter = AccountPropertyTransaction.createAddressFilter(
+        const addressPropertyFilter = AccountPropertyModification.createForAddress(
             PropertyModificationType.Add,
             address,
         );
@@ -609,7 +629,7 @@ describe('TransactionMapping - createFromDTO (Transaction.toJSON() feed)', () =>
 
     it('should create AccountPropertyMosaicTransaction', () => {
         const mosaicId = new MosaicId([2262289484, 3405110546]);
-        const mosaicPropertyFilter = AccountPropertyTransaction.createMosaicFilter(
+        const mosaicPropertyFilter = AccountPropertyModification.createForMosaic(
             PropertyModificationType.Add,
             mosaicId,
         );
@@ -630,7 +650,7 @@ describe('TransactionMapping - createFromDTO (Transaction.toJSON() feed)', () =>
 
     it('should create AccountPropertyMosaicTransaction', () => {
         const entityType = TransactionType.ADDRESS_ALIAS;
-        const entityTypePropertyFilter = AccountPropertyTransaction.createEntityTypeFilter(
+        const entityTypePropertyFilter = AccountPropertyModification.createForEntityType(
             PropertyModificationType.Add,
             entityType,
         );